added moderation and fixed sar
