+++ src/cogs.txt
@@ -3,3 +3,4 @@ contributors
 code
 git
 fun
+moderation
